+++ yii2-adapter/routes/web.php
@@ -4,11 +4,24 @@ use CraftCms\Cms\Cms;
 use CraftCms\Cms\Http\Middleware\RequireAdminChanges;
 use CraftCms\Cms\Route\Routes as CraftRoutes;
 use CraftCms\Yii2Adapter\Http\ExcludeCsrfValidationForLegacyController;
+use CraftCms\Yii2Adapter\Http\FallbackTransformController;
 use CraftCms\Yii2Adapter\Http\LegacyMiddleware;
 use CraftCms\Yii2Adapter\Http\SavePluginSettingsController;
 use Illuminate\Support\Facades\Route;
 
 $routes = app(CraftRoutes::class);
+$sharedActionRouteGroups = $routes->actionTriggerRoutePrefix() === $routes->cpActionTriggerRoutePrefix()
+    ? [[$routes->cpActionTriggerRoutePrefix(), ['craft.cp']]]
+    : [
+        [$routes->actionTriggerRoutePrefix(), ['craft.web']],
+        [$routes->cpActionTriggerRoutePrefix(), ['craft.cp']],
+    ];
+
+foreach ($sharedActionRouteGroups as [$prefix, $middleware]) {
+    Route::middleware(['web', 'craft', ...$middleware])
+        ->prefix($prefix)
+        ->get('assets/generate-fallback-transform', FallbackTransformController::class);
+}
 
 Route::middleware([
     'web',